Fixes slow smooth scroll speeds
It takes too long to use the TOC links on the netlify docs site. Personally, the experience of watching a page scroll for over a second makes me dizzy. That is probably a stronger-than-typical reaction. The goal of this change is to set a 'ceiling' for how long a smooth scroll animation may last. `smooth-scroll` uses a default speed of 1000px scrolled every 300 milliseconds. By applying a ceiling, longer page jumps will scroll faster.

See also: https://www.npmjs.com/package/smooth-scroll#options-and-settings